+++ src/hooks/use-vcard-saver.ts
@@ -1,0 +1,7 @@
+export const useVCardSaver = (onSave?: () => void) => {
+  const handleSaveVCard = () => {
+    if (onSave) onSave();
+  };
+
+  return { handleSaveVCard };
+};
